pyproject.toml
[project]
name = "mini-project-2-ecommerce-backend"
version = "0.1.0"
description = "Minimal Django + DRF e-commerce backend (inventory, cart, order) with Celery+Redis."
requires-python = ">=3.11"
dependencies = [
  "Django==5.2.13",
  "djangorestframework==3.17.1",
  "celery==5.6.3",
  "redis==7.4.0",
  "django-celery-results==2.6.0",
]

[dependency-groups]
dev = []

